Refine Analyzer and LogicalPlanner
